fix: move inline tests to module_access/tests.rs to resolve E0428 duplicate name
The include!() composition pattern places all submodule contents in the
same namespace, so the inline `mod tests` in helpers.rs conflicted with
module_parsing.rs's `mod tests` declaration.
